+++ errors.go
@@ -2,7 +2,7 @@ package servicebus
 
 import (
 	"fmt"
-	"github.com/Azure/azure-amqp-common-go/rpc"
+	"github.com/Azure/azure-amqp-common-go/v2/rpc"
 	"reflect"
 )