+++ sender.go
@@ -27,9 +27,8 @@ import (
 	"math/rand"
 	"time"
 
-	"github.com/Azure/azure-amqp-common-go/log"
-	"github.com/Azure/azure-amqp-common-go/uuid"
-	"github.com/opentracing/opentracing-go"
+	"github.com/Azure/azure-amqp-common-go/v2/uuid"
+	"github.com/devigned/tab"
 	"pack.ag/amqp"
 )
 
@@ -50,6 +49,8 @@ type (
 
 	eventer interface {
 		toMsg() (*amqp.Message, error)
+		GetKeyValues() map[string]interface{}
+		Set(key string, value interface{})
 	}
 
 	// SenderOption provides a way to customize a Sender
@@ -58,8 +59,8 @@ type (
 
 // NewSender creates a new Service Bus message Sender given an AMQP client and entity path
 func (ns *Namespace) NewSender(ctx context.Context, entityPath string, opts ...SenderOption) (*Sender, error) {
-	span, ctx := ns.startSpanFromContext(ctx, "sb.Sender.NewSender")
-	defer span.Finish()
+	ctx, span := ns.startSpanFromContext(ctx, "sb.Namespace.NewSender")
+	defer span.End()
 
 	s := &Sender{
 		namespace:  ns,
@@ -68,26 +69,26 @@ func (ns *Namespace) NewSender(ctx context.Context, entityPath string, opts ...S
 
 	for _, opt := range opts {
 		if err := opt(s); err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return nil, err
 		}
 	}
 
 	err := s.newSessionAndLink(ctx)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 	}
 	return s, err
 }
 
 // Recover will attempt to close the current session and link, then rebuild them
 func (s *Sender) Recover(ctx context.Context) error {
-	span, ctx := s.startProducerSpanFromContext(ctx, "sb.Sender.Recover")
-	defer span.Finish()
+	ctx, span := s.startProducerSpanFromContext(ctx, "sb.Sender.Recover")
+	defer span.End()
 
 	// we expect the Sender, session or client is in an error state, ignore errors
 	closeCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	closeCtx = opentracing.ContextWithSpan(closeCtx, span)
+	closeCtx = tab.NewContext(closeCtx, span)
 	defer cancel()
 	_ = s.sender.Close(closeCtx)
 	_ = s.session.Close(closeCtx)
@@ -97,8 +98,8 @@ func (s *Sender) Recover(ctx context.Context) error {
 
 // Close will close the AMQP connection, session and link of the Sender
 func (s *Sender) Close(ctx context.Context) error {
-	span, _ := s.startProducerSpanFromContext(ctx, "sb.Sender.Close")
-	defer span.Finish()
+	ctx, span := s.startProducerSpanFromContext(ctx, "sb.Sender.Close")
+	defer span.End()
 
 	err := s.sender.Close(ctx)
 	if err != nil {
@@ -120,8 +121,8 @@ func (s *Sender) Close(ctx context.Context) error {
 //
 // This will retry sending the message if the server responds with a busy error.
 func (s *Sender) Send(ctx context.Context, msg *Message, opts ...SendOption) error {
-	span, ctx := s.startProducerSpanFromContext(ctx, "sb.Sender.Send")
-	defer span.Finish()
+	ctx, span := s.startProducerSpanFromContext(ctx, "sb.Sender.Send")
+	defer span.End()
 
 	if msg.SessionID == nil {
 		msg.SessionID = &s.session.SessionID
@@ -132,7 +133,7 @@ func (s *Sender) Send(ctx context.Context, msg *Message, opts ...SendOption) err
 	if msg.ID == "" {
 		id, err := uuid.NewV4()
 		if err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return err
 		}
 		msg.ID = id.String()
@@ -141,7 +142,7 @@ func (s *Sender) Send(ctx context.Context, msg *Message, opts ...SendOption) err
 	for _, opt := range opts {
 		err := opt(msg)
 		if err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return err
 		}
 	}
@@ -150,30 +151,30 @@ func (s *Sender) Send(ctx context.Context, msg *Message, opts ...SendOption) err
 }
 
 func (s *Sender) trySend(ctx context.Context, evt eventer) error {
-	sp, ctx := s.startProducerSpanFromContext(ctx, "sb.Sender.trySend")
-	defer sp.Finish()
+	ctx, sp := s.startProducerSpanFromContext(ctx, "sb.Sender.trySend")
+	defer sp.End()
 
-	err := opentracing.GlobalTracer().Inject(sp.Context(), opentracing.TextMap, evt)
+	err := sp.Inject(evt)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 
 	msg, err := evt.toMsg()
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 
 	if msg.Properties != nil {
-		sp.SetTag("sb.message-id", msg.Properties.MessageID)
+		sp.AddAttributes(tab.StringAttribute("sb.message.id", msg.Properties.MessageID.(string)))
 	}
 
 	for {
 		select {
 		case <-ctx.Done():
 			if ctx.Err() != nil {
-				log.For(ctx).Error(err)
+				tab.For(ctx).Error(err)
 			}
 			return ctx.Err()
 		default:
@@ -186,16 +187,16 @@ func (s *Sender) trySend(ctx context.Context, evt eventer) error {
 
 			switch err.(type) {
 			case *amqp.Error, *amqp.DetachError:
-				log.For(ctx).Debug("amqp error, delaying 4 seconds: " + err.Error())
+				tab.For(ctx).Debug("amqp error, delaying 4 seconds: " + err.Error())
 				skew := time.Duration(rand.Intn(1000)-500) * time.Millisecond
 				time.Sleep(4*time.Second + skew)
 				err := s.Recover(ctx)
 				if err != nil {
-					log.For(ctx).Debug("failed to recover connection")
+					tab.For(ctx).Debug("failed to recover connection")
 				}
-				log.For(ctx).Debug("recovered connection")
+				tab.For(ctx).Debug("recovered connection")
 			default:
-				log.For(ctx).Error(err)
+				tab.For(ctx).Error(err)
 				return err
 			}
 		}
@@ -216,25 +217,25 @@ func (s *Sender) getFullIdentifier() string {
 
 // newSessionAndLink will replace the existing session and link
 func (s *Sender) newSessionAndLink(ctx context.Context) error {
-	span, ctx := s.startProducerSpanFromContext(ctx, "sb.Sender.newSessionAndLink")
-	defer span.Finish()
+	ctx, span := s.startProducerSpanFromContext(ctx, "sb.Sender.newSessionAndLink")
+	defer span.End()
 
 	connection, err := s.namespace.newConnection()
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 	s.connection = connection
 
 	err = s.namespace.negotiateClaim(ctx, connection, s.getAddress())
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 
 	amqpSession, err := connection.NewSession()
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 
@@ -242,13 +243,13 @@ func (s *Sender) newSessionAndLink(ctx context.Context) error {
 		amqp.LinkSenderSettle(amqp.ModeUnsettled),
 		amqp.LinkTargetAddress(s.getAddress()))
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 
 	s.session, err = newSession(amqpSession)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 	if s.sessionID != nil {