+++ topic.go
@@ -28,9 +28,9 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/Azure/azure-amqp-common-go/log"
-	"github.com/Azure/azure-amqp-common-go/uuid"
+	"github.com/Azure/azure-amqp-common-go/v2/uuid"
 	"github.com/Azure/go-autorest/autorest/date"
+	"github.com/devigned/tab"
 )
 
 type (
@@ -96,12 +96,12 @@ func (ns *Namespace) NewTopic(name string, opts ...TopicOption) (*Topic, error) 
 
 // Send sends messages to the Topic
 func (t *Topic) Send(ctx context.Context, event *Message, opts ...SendOption) error {
-	span, ctx := t.startSpanFromContext(ctx, "sb.Topic.Send")
-	defer span.Finish()
+	ctx, span := t.startSpanFromContext(ctx, "sb.Topic.Send")
+	defer span.End()
 
 	err := t.ensureSender(ctx)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 	return t.sender.Send(ctx, event, opts...)
@@ -109,19 +109,19 @@ func (t *Topic) Send(ctx context.Context, event *Message, opts ...SendOption) er
 
 // SendBatch sends a batch of messages to the Topic
 func (t *Topic) SendBatch(ctx context.Context, iterator BatchIterator) error {
-	span, ctx := t.startSpanFromContext(ctx, "sb.Topic.SendBatch")
-	defer span.Finish()
+	ctx, span := t.startSpanFromContext(ctx, "sb.Topic.SendBatch")
+	defer span.End()
 
 	err := t.ensureSender(ctx)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 
 	for !iterator.Done() {
 		id, err := uuid.NewV4()
 		if err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return err
 		}
 
@@ -129,12 +129,12 @@ func (t *Topic) SendBatch(ctx context.Context, iterator BatchIterator) error {
 			SessionID: t.sender.sessionID,
 		})
 		if err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return err
 		}
 
 		if err := t.sender.trySend(ctx, batch); err != nil {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return err
 		}
 	}
@@ -159,13 +159,13 @@ func (t *Topic) NewSender(ctx context.Context, opts ...SenderOption) (*Sender, e
 
 // Close the underlying connection to Service Bus
 func (t *Topic) Close(ctx context.Context) error {
-	span, ctx := t.startSpanFromContext(ctx, "sb.Topic.Close")
-	defer span.Finish()
+	ctx, span := t.startSpanFromContext(ctx, "sb.Topic.Close")
+	defer span.End()
 
 	if t.sender != nil {
 		err := t.sender.Close(ctx)
 		if err != nil && !isConnectionClosed(err) {
-			log.For(ctx).Error(err)
+			tab.For(ctx).Error(err)
 			return err
 		}
 	}
@@ -190,16 +190,16 @@ func (t *Topic) NewTransferDeadLetter() *TransferDeadLetter {
 //   - The destination queue or topic is disabled or deleted.
 //   - The destination queue or topic exceeds the maximum entity size.
 func (t *Topic) NewTransferDeadLetterReceiver(ctx context.Context, opts ...ReceiverOption) (ReceiveOner, error) {
-	span, ctx := t.startSpanFromContext(ctx, "sb.Topic.NewTransferDeadLetterReceiver")
-	defer span.Finish()
+	ctx, span := t.startSpanFromContext(ctx, "sb.Topic.NewTransferDeadLetterReceiver")
+	defer span.End()
 
 	transferDeadLetterEntityPath := strings.Join([]string{t.Name, TransferDeadLetterQueueName}, "/")
 	return t.namespace.NewReceiver(ctx, transferDeadLetterEntityPath, opts...)
 }
 
 func (t *Topic) ensureSender(ctx context.Context) error {
-	span, ctx := t.startSpanFromContext(ctx, "sb.Topic.ensureSender")
-	defer span.Finish()
+	ctx, span := t.startSpanFromContext(ctx, "sb.Topic.ensureSender")
+	defer span.End()
 
 	t.senderMu.Lock()
 	defer t.senderMu.Unlock()
@@ -210,7 +210,7 @@ func (t *Topic) ensureSender(ctx context.Context) error {
 
 	s, err := t.namespace.NewSender(ctx, t.Name)
 	if err != nil {
-		log.For(ctx).Error(err)
+		tab.For(ctx).Error(err)
 		return err
 	}
 	t.sender = s